Removed unnecessary flow
